% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bigelow_card}
\alias{bigelow_card}
\title{Creates and returns a Bigelow-style plot card with optional header, main content, 
and optional footer}
\usage{
bigelow_card(..., headerContent = NULL, footerContent = NULL)
}
\arguments{
\item{...}{content to put in main body, typically a plot}

\item{headerContent}{content, text to put in header}

\item{footerContent}{content, text to put in footer}
}
\description{
Creates and returns a Bigelow-style plot card with optional header, main content, 
and optional footer
}
