% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bigelow_header}
\alias{bigelow_header}
\alias{bigelow_main_body}
\alias{bigelow_footer}
\title{Creates and returns a Bigelow-style header for an R Shiny application.}
\usage{
bigelow_header(left_hand, right_hand = NULL)

bigelow_main_body(...)

bigelow_footer(left_hand)
}
\arguments{
\item{left_hand, }{div, material to have on left hand side of footer.
If this is pure text it is parsed to a header element}

\item{right_hand, }{div, material to have on right hand side of footer or NULL}

\item{..., }{content}
}
\value{
div, footer element with bigelow logo.

div, content with wrapped styling

div, footer element with bigelow logo.
}
\description{
Creates and returns a Bigelow-style header for an R Shiny application.

Creates and returns a Bigelow-style main body for an R Shiny application.
Basically just gives the main content a slight padding & restricts max width.

Creates and returns a Bigelow-style footer for an R Shiny application.
Required to have bigelow_logo.svg in the www/images folder
}
