% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bigelow_style_plots}
\alias{bigelow_style_plots}
\title{Applies bigelow-style theme colors to \strong{ggplot2}, \strong{base}, and \strong{lattice} graphics within an R Shiny application. Wraps \code{\link[thematic]{thematic_shiny}}.}
\usage{
bigelow_style_plots(
  accent = "#02A5DD",
  sequential = viridis::viridis(4),
  qualitative = thematic::okabe_ito()
)
}
\arguments{
\item{accent}{a color for making certain graphical markers 'stand out', i.e. fitted line color in \code{\link[ggplot2]{geom_smooth}}. Default color is Bigelow sky blue, replace if high contrast required.}

\item{sequential}{a color palette for graphical markers encoding numeric values. Can be a vector of color codes or a \code{\link[thematic]{sequential_gradient}} object. Defaults to \code{\link[viridis]{viridis}} color scheme, which is colorblind safe.}

\item{qualitative}{a color palette for graphical markers that encode qualitative values. Defaults to \code{\link[thematic]{okabe_ito}}, which is verified to be colorblind safe.}
}
\value{
global theme object.
}
\description{
Applies bigelow-style theme colors to \strong{ggplot2}, \strong{base}, and \strong{lattice} graphics within an R Shiny application. Wraps \code{\link[thematic]{thematic_shiny}}.
}
\examples{
\dontrun{
bigelow_style_plots()
shinyApp(ui, server)
}
}
