% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bigelow_theme}
\alias{bigelow_theme}
\title{Creates and returns a Bigelow-style theme object for R Shiny.}
\usage{
bigelow_theme()
}
\value{
a Shiny application \strong{bslib} theme object with Bigelow-style theming.
}
\description{
Creates and returns a Bigelow-style theme object for R Shiny.
}
\examples{
\dontrun{
ui <- fluidPage(
 theme = bigelow_theme(),
 p("Hello World!"))
server <- function(input, output, session) {}
shinyApp(ui, server)
}
}
