% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigelowshinytheme-package.R
\docType{package}
\name{bigelowshinytheme-package}
\alias{bigelowshinytheme-package}
\alias{bigelowshinytheme}
\title{bigelowshinytheme}
\description{
Package encompassing a theme object and supporting structure function for Shiny apps matching official Bigelow Laboratory styling.
}
\details{
Bigelow-style theming for Shiny applications
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/bigelowshinytheme}
}

}
\author{
\strong{Maintainer}: Johnson Omi \email{ojohnson@bigelow.org}

}
\keyword{internal}
