% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_www.R
\name{copy_www}
\alias{copy_www}
\title{Copies www folder with supporting images to specified directory}
\usage{
copy_www(dest = getwd())
}
\arguments{
\item{dest}{char, path to intended parent directory of www folder}
}
\value{
results from file.copy
}
\description{
Copies www folder with supporting images to specified directory
}
